clearly-defined: Move enums and errors to separate files
To reduce the size of the ClearlyDefinedService class a bit for a better
overview.
