+++ clearly-defined/src/main/kotlin/Errors.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2020 Bosch.IO GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.ossreviewtoolkit.clearlydefined
+
+/**
+ * See https://github.com/clearlydefined/service/blob/c47a989/app.js#L201-L205.
+ */
+data class ErrorResponse(
+    val error: Error
+)
+
+data class Error(
+    val code: String,
+    val message: String,
+    val innererror: InnerError
+)
+
+data class InnerError(
+    val name: String,
+    val message: String,
+    val stack: String
+)
